moved the champion cost filter into it's own component and imported it into the champions page
